# tsam-golang/docs/swagger.yaml
basePath: /api/v1/tsam
definitions:
  errors.HTTPError:
    properties:
      errorKey:
        example: unable to fetch data
        type: string
      httpstatus:
        example: 400
        type: integer
    type: object
  errors.ValidationError:
    properties:
      error:
        example: Name must be specified
        type: string
    type: object
...
  model.Channel:
    properties:
      channelName:
        example: Golang Developer
        type: string
      id:
        type: string
    type: object
...
  model.Discussion:
    properties:
      channel:
        $ref: '#/definitions/model.Channel'
        type: object
      description:
        example: Question Description
        type: string
      id:
        type: string
      isSolve:
        example: true
        type: boolean
      question:
        example: Question string
        type: string
      talent:
        $ref: '#/definitions/model.Talent'
        type: object
    type: object
...
  model.Reply:
    properties:
      bestReply:
        example: true
        type: boolean
      comment:
        items:
          $ref: '#/definitions/model.Reply'
        type: array
      discussionID:
        example: cfe25758-f5fe-48f0-874d-e72cd4edd9b9
        type: string
      id:
        type: string
      replier:
        $ref: '#/definitions/model.Talent'
        type: object
      reply:
        example: Answer Against Question
        type: string
      replyID:
        example: cfe25758-f5fe-48f0-874d-e72cd4edd9b9
        type: string
    type: object
  model.ReplyDTO:
    properties:
      discussionDTO:
        $ref: '#/definitions/model.Discussion'
        type: object
      replyDTO:
        items:
          $ref: '#/definitions/model.Reply'
        type: array
    type: object
...
  model.Talent:
    properties:
      id:
        type: string
    type: object
...
host: swabhav-tsam.herokuapp.com
info:
  contact:
    email: info@swabhavtechlabs.com
    name: Swabhav Techlabs
  description: Short Description
  license:
    name: Apache 2.0
  title: Talent Sourcing and Mentoring (TSAM)
  version: "1.0"
paths:
...
  /reply:
    post:
      consumes:
      - application/json
      description: Add New Reply
      parameters:
      - description: Add Reply
        in: body
        name: reply
        required: true
        schema:
          $ref: '#/definitions/model.Reply'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: plain
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      tags:
      - community-forum
  /reply/{replyid}:
    delete:
      consumes:
      - application/json
      description: Delete Reply By Reply ID
      parameters:
      - description: Reply ID For Delete Reply
        format: uuid.UUID
        in: path
        name: replyid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: plain
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      tags:
      - community-forum
    get:
      consumes:
      - application/json
      description: GetReply Reply By Reply ID
      parameters:
      - description: Reply ID For Get Reply
        format: uuid.UUID
        in: path
        name: replyid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Reply'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      tags:
      - community-forum
    put:
      consumes:
      - application/json
      description: Update Reply By Reply ID
      parameters:
      - description: Add Reply
        in: body
        name: reply
        required: true
        schema:
          $ref: '#/definitions/model.Reply'
      - description: Reply ID For Update Reply
        format: uuid.UUID
        in: path
        name: replyid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: plain
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      tags:
      - community-forum
  /reply/discussion/{discussionid}/{limit}/{offset}:
    get:
      consumes:
      - application/json
      description: Return multiple Reply By Discussion ID
      parameters:
      - description: total number of result
        format: int
        in: path
        name: limit
        required: true
        type: integer
      - description: page number
        format: int
        in: path
        name: offset
        required: true
        type: integer
      - description: Get ReplyDTO By Discussion id
        format: uuid.UUID
        in: path
        name: discussionid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/model.ReplyDTO'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      tags:
      - community-forum
...
swagger: "2.0"